Mise a jour

+++ script_Red-Hat.sh
@@ -6,17 +6,22 @@ echo "Que voulez mettre a jour:"
 echo " 1.Update système(dnf)"
 echo " 2.Update Faltpak"
 echo " 3.Update système et Faltpak(dnf+faltpak)"
-echo " 4.Update système et Faltpak avec redemarage(dnf+faltpak+reboot)"
-echo " 5.Grub"
-echo " 6.Lister les Kernel"
+echo " 4.Grub"
 echo " "
-echo "Que voulez-vous faire(1,2,3,...):"
+echo "[1,2,3,4]"
 read reponse
 
 #Update du systeme 
 if [ "$reponse" == "1" ]
 then
     dnf update -y
+    echo "Voulez-vous redemarer ?"
+    echo "[O,N]"
+    read reponse2
+    if ["$reponse2" == "o" ,"$reponse2" == "O" ]
+    then
+    reboot
+    fi
 fi
 #___________________
 
@@ -32,28 +37,27 @@ if [ "$reponse" == "3" ]
 then
     dnf update -y
     flatpak update
+    echo "Voulez-vous redemarer ?"
+    echo "[O,N]"
+    read reponse2
+    if ["$reponse2" == "o" ,"$reponse2" == "O" ]
+    then
+    reboot
+    fi
 fi
 #___________________
 
-#Update du systeme + flatpak + reboot
-if [ "$reponse" == "4" ]
-then
-    dnf update -y
-    flatpak update
-    reboot
-fi
-#___________________
 
 #Update du grub
-if [ "$reponse" == "5" ]
+if [ "$reponse" == "4" ]
 then
     grub2-mkconfig -o /boot/grub2/grub.cfg
-fi
-#___________________
-
-#liste des kernel disponible
-if [ "$reponse" == "6" ]
-then
-   dnf list kernel
+    echo "Voulez-vous redemarer ?"
+    echo "[O,N]"
+    read reponse2
+    if ["$reponse2" == "o" ,"$reponse2" == "O" ]
+    then
+    reboot
+    fi
 fi
 #___________________
